refactor: convert RoleMemberOption to base component for both use cases
